syd admin page

+++ Group12_iCAREAPP/Controllers/iCAREUsersController.cs
@@ -39,30 +39,69 @@ namespace Group12_iCAREAPP.Controllers
         // GET: iCAREUsers/Create
         public ActionResult Create()
         {
-            ViewBag.ID = new SelectList(db.iCAREAdmin, "ID", "ID");
-            ViewBag.passwordID = new SelectList(db.UserPassword, "ID", "password");
-            ViewBag.ID = new SelectList(db.iCAREWorker, "ID", "profession");
+            ViewBag.creatorID = new SelectList(db.iCAREAdmin, "ID", "ID");
+            ViewBag.roleID = new SelectList(db.iCAREWorker, "ID", "profession");
             return View();
         }
 
         // POST: iCAREUsers/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "ID,name,passwordID")] iCAREUser iCAREUser)
+        public ActionResult Create([Bind(Include = "ID,name,password,passwordID,profession,creatorID,roleID")] iCAREUserViewModel viewModel)
         {
             if (ModelState.IsValid)
             {
-                db.iCAREUser.Add(iCAREUser);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                using (var transaction = db.Database.BeginTransaction())
+                {
+                    try
+                    {
+                        //create the UserPassword
+                        var userPassword = new UserPassword
+                        {
+                            password = viewModel.password,
+                            ID = viewModel.passwordID
+                        };
+
+                        db.UserPassword.Add(userPassword);
+                        db.SaveChanges();
+
+                        //create the iCAREUser
+                        var iCAREUser = new iCAREUser
+                        {
+                            ID = viewModel.ID,
+                            name = viewModel.name,
+                            passwordID = userPassword.ID
+                        };
+
+                        db.iCAREUser.Add(iCAREUser);
+                        db.SaveChanges();
+
+                        //create the iCAREWorker
+                        var iCAREWorker = new iCAREWorker
+                        {
+                            ID = iCAREUser.ID,
+                            profession = viewModel.profession,
+                            creatorID = viewModel.creatorID,
+                            roleID = viewModel.roleID
+                        };
+
+                        db.iCAREWorker.Add(iCAREWorker);
+                        db.SaveChanges();
+
+                        transaction.Commit();
+                        return RedirectToAction("Index");
+                    }
+                    catch (Exception)
+                    {
+                        transaction.Rollback();
+                        ModelState.AddModelError("", "An error occurred while creating the user and worker.");
+                    }
+                }
             }
 
-            ViewBag.ID = new SelectList(db.iCAREAdmin, "ID", "ID", iCAREUser.ID);
-            ViewBag.passwordID = new SelectList(db.UserPassword, "ID", "password", iCAREUser.passwordID);
-            ViewBag.ID = new SelectList(db.iCAREWorker, "ID", "profession", iCAREUser.ID);
-            return View(iCAREUser);
+            ViewBag.creatorID = new SelectList(db.iCAREAdmin, "ID", "ID", viewModel.creatorID);
+            ViewBag.roleID = new SelectList(db.iCAREWorker, "ID", "profession", viewModel.roleID);
+            return View(viewModel);
         }
 
         // GET: iCAREUsers/Edit/5
@@ -72,35 +111,100 @@ namespace Group12_iCAREAPP.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            iCAREUser iCAREUser = db.iCAREUser.Find(id);
+
+            var iCAREUser = db.iCAREUser.Find(id);
             if (iCAREUser == null)
             {
                 return HttpNotFound();
             }
-            ViewBag.ID = new SelectList(db.iCAREAdmin, "ID", "ID", iCAREUser.ID);
-            ViewBag.passwordID = new SelectList(db.UserPassword, "ID", "password", iCAREUser.passwordID);
-            ViewBag.ID = new SelectList(db.iCAREWorker, "ID", "profession", iCAREUser.ID);
-            return View(iCAREUser);
+
+            var userPassword = db.UserPassword.Find(iCAREUser.passwordID);
+            var iCAREWorker = db.iCAREWorker.Find(iCAREUser.ID);
+
+            var viewModel = new iCAREUserViewModel
+            {
+                ID = iCAREUser.ID,
+                name = iCAREUser.name,
+                password = userPassword?.password,
+                passwordID = iCAREUser.passwordID,
+                profession = iCAREWorker?.profession,
+                creatorID = iCAREWorker?.creatorID,
+                roleID = iCAREWorker?.roleID
+            };
+
+            // Get distinct roles
+            var distinctRoles = db.iCAREWorker.Select(w => w.profession).Distinct().ToList();
+
+            ViewBag.creatorID = new SelectList(db.iCAREAdmin, "ID", "ID", viewModel.creatorID);
+            ViewBag.roleID = new SelectList(distinctRoles, viewModel.roleID);
+            return View(viewModel);
         }
+
+
 
         // POST: iCAREUsers/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ID,name,passwordID")] iCAREUser iCAREUser)
+        public ActionResult Edit([Bind(Include = "ID,name,password,passwordID,profession,creatorID,roleID")] iCAREUserViewModel viewModel)
         {
             if (ModelState.IsValid)
             {
-                db.Entry(iCAREUser).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                using (var transaction = db.Database.BeginTransaction())
+                {
+                    try
+                    {
+                        //find the existing iCAREUser
+                        var iCAREUser = db.iCAREUser.Find(viewModel.ID);
+                        if (iCAREUser == null)
+                        {
+                            return HttpNotFound();
+                        }
+
+                        //update the UserPassword
+                        var userPassword = db.UserPassword.Find(iCAREUser.passwordID);
+                        if (userPassword != null)
+                        {
+                            userPassword.password = viewModel.password;
+                            db.Entry(userPassword).State = EntityState.Modified;
+                        }
+
+                        //update the iCAREUser
+                        iCAREUser.name = viewModel.name;
+                        db.Entry(iCAREUser).State = EntityState.Modified;
+
+                        //update the iCAREWorker
+                        var iCAREWorker = db.iCAREWorker.Find(iCAREUser.ID);
+                        if (iCAREWorker != null)
+                        {
+                            iCAREWorker.profession = viewModel.profession;
+                            iCAREWorker.creatorID = viewModel.creatorID;
+                            iCAREWorker.roleID = viewModel.roleID;
+                            db.Entry(iCAREWorker).State = EntityState.Modified;
+                        }
+
+                        //save changes
+                        db.SaveChanges();
+
+                        //commit transaction
+                        transaction.Commit();
+                        return RedirectToAction("Index");
+                    }
+                    catch (Exception)
+                    {
+                        //if any error occurs
+                        transaction.Rollback();
+                        ModelState.AddModelError("", "An error occurred while updating the user and associated records.");
+                    }
+                }
             }
-            ViewBag.ID = new SelectList(db.iCAREAdmin, "ID", "ID", iCAREUser.ID);
-            ViewBag.passwordID = new SelectList(db.UserPassword, "ID", "password", iCAREUser.passwordID);
-            ViewBag.ID = new SelectList(db.iCAREWorker, "ID", "profession", iCAREUser.ID);
-            return View(iCAREUser);
+
+            ViewBag.creatorID = new SelectList(db.iCAREAdmin, "ID", "ID", viewModel.creatorID);
+            ViewBag.roleID = new SelectList(db.iCAREWorker, "ID", "profession", viewModel.roleID);
+            return View(viewModel);
         }
+
 
         // GET: iCAREUsers/Delete/5
         public ActionResult Delete(string id)
@@ -122,11 +226,52 @@ namespace Group12_iCAREAPP.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(string id)
         {
-            iCAREUser iCAREUser = db.iCAREUser.Find(id);
-            db.iCAREUser.Remove(iCAREUser);
-            db.SaveChanges();
+            using (var transaction = db.Database.BeginTransaction())
+            {
+                try
+                {
+                    //find the iCAREUser
+                    iCAREUser iCAREUser = db.iCAREUser.Find(id);
+                    if (iCAREUser == null)
+                    {
+                        return HttpNotFound();
+                    }
+
+                    //find the associated iCAREWorker
+                    iCAREWorker iCAREWorker = db.iCAREWorker.Find(id);
+                    if (iCAREWorker != null)
+                    {
+                        db.iCAREWorker.Remove(iCAREWorker);
+                    }
+
+                    //find the associated UserPassword
+                    UserPassword userPassword = db.UserPassword.Find(iCAREUser.passwordID);
+                    if (userPassword != null)
+                    {
+                        db.UserPassword.Remove(userPassword);
+                    }
+
+                    //remove the iCAREUser
+                    db.iCAREUser.Remove(iCAREUser);
+
+                    //save changes
+                    db.SaveChanges();
+
+                    //commit transaction
+                    transaction.Commit();
+                }
+                catch (Exception)
+                {
+                    //if any error occurs
+                    transaction.Rollback();
+                    ModelState.AddModelError("", "An error occurred while deleting the user and associated records.");
+                    return RedirectToAction("Delete", new { id });
+                }
+            }
+
             return RedirectToAction("Index");
         }
+
 
         protected override void Dispose(bool disposing)
         {
